+++ extension/core/ELanguage/ELanguage.js
@@ -2,12 +2,9 @@
 import languageData from "./lang/lang.json" with { type: "json" };
 import AutoLang from './AutoLanguage.js'
 
-const langISOCode = AutoLang()
-const jsonObj = languageData
-const langObj = jsonObj[langISOCode] ? jsonObj[langISOCode]: jsonObj["en"];
+export function getLang(langISOCode) {
+    const jsonObj = languageData
+    const langObj = jsonObj[langISOCode] ? jsonObj[langISOCode]: jsonObj["en"];
+    return langObj
+}
 
-//set html lang attr to readed iso
-// document.documentElement.lang = langISOCode;
-
-export const getLang = () => langObj;
-export const getLangVar = (key) => langObj[key]